LMS/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyparsing==3.2.3
tzdata==2024.2